Added PostsPage component to App.js and moved all app state and methods there
